2/3/16 - Revised code with call functions. 2/9/16 - Control layer classes. 2/11/16 - Committing changes. 2/11/16 - Corrected errors (!=). 2/11/16 - Added first two test cases. 2/11/16 - Added remaining test cases. 2/11/16 - Minor edits. 2/11/16 - More edits to try to clear errors. 2/11/16 - Final edits.  100% pass rate. 2/12/16 - Adding PuzzleSceneWater class and test packages. 2/13/16 - Final edits to PuzzleSceneWater.  100% pass rate. 2/18/16 - MainMenuView additions. 2/19/16 - Added HelpMenuView. 2/19/16 - Minor edits. 2/23/16 - Added Items Menu. 2/24/16 - Attempted Water Puzzle View 03/03/16 - Extended View Interface 03/05/16 - Added SaveGame.java 03/05/16 - Major code corrections with Brother Jones.

+++ theHunt2/test/byu/cit260/theHunt2/control/PuzzleSceneWaterTest.java
@@ -47,7 +47,7 @@ public class PuzzleSceneWaterTest {
         PuzzleSceneWater instance = new PuzzleSceneWater();
         
         //call function to run test
-        double result = instance.calcTubFillRate(volume, gpm);
+        double result = instance.calcTubFillRate((int) volume, gpm);
         
         //compare expected return value with actual value returned
         assertEquals(expResult, result, 18.18);
@@ -81,7 +81,7 @@ public class PuzzleSceneWaterTest {
         expResult = -1; // expected output returned value
         
         // call function to run test
-        result = instance.calcTubFillRate( volume, gpm);
+        result = instance.calcTubFillRate((int) volume, gpm);
         
         // compare expected return value with actual value returned
         assertEquals(expResult, result, -1.0);
@@ -100,7 +100,7 @@ public class PuzzleSceneWaterTest {
         expResult = -1; // expected output returned value
         
         // call function to run test
-        result = instance.calcTubFillRate( volume, gpm);
+        result = instance.calcTubFillRate((int) volume, gpm);
         
         // compare expected return value with actual value returned
         assertEquals(expResult, result, -1.0);
@@ -119,7 +119,7 @@ public class PuzzleSceneWaterTest {
         expResult = 18.18181818181818; // expected output returned value
         
         // call function to run test
-        result = instance.calcTubFillRate( volume, gpm);
+        result = instance.calcTubFillRate((int) volume, gpm);
         
         // compare expected return value with actual value returned
         assertEquals(expResult, result, 18.18181818181818);
@@ -138,7 +138,7 @@ public class PuzzleSceneWaterTest {
         expResult = 54.54; // expected output returned value
         
         // call function to run test
-        result = instance.calcTubFillRate( volume, gpm);
+        result = instance.calcTubFillRate((int) volume, gpm);
         
         // compare expected return value with actual value returned
         assertEquals(expResult, result, 54.54);
